+++ poem/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `X-Accel-Buffering: no` header to SSE response.
 - Add `AutoCertBuilder::contact` method to add a contact email for ACME account.
+- Add `TestJsonArray::assert_is_empty` and `TestJsonObject::assert_is_empty` methods.
 
 # [1.3.4] 2022-2-21